# aws/corehandlers: speed up validation of []byte