Request validation logic ensures that struct fields have appropriate values.
When validating slices and maps, the goal is to find any structs they may
contain, since only struct fields have validation rules. Slices of uint8,
aka []byte, do not lead to structs and so cannot cause validation to fail.
Speed up validation of this (unnamed) type.

This is particularly important for APIs such as Kinesis which can contain
byte slices for megabytes of data - each byte does not need to be validated
individually.

The below output of golang.org/x/tools/cmd/benchcmp is for a Kinesis
PutRecords request including 100 Records of 10kB each.

benchmark                  old ns/op     new ns/op     delta
BenchmarkValidateAny-8     310052182     216116        -99.93%

benchmark                  old allocs     new allocs     delta
BenchmarkValidateAny-8     3002424        1909           -99.94%

benchmark                  old bytes     new bytes     delta
BenchmarkValidateAny-8     48079408      19274         -99.96%