@jasdel 2016-02-05
Looks really good, thanks for taking the time to put together this PR.  The only comment I have is a minor tweak to the method used to determine if the value is a `[]byte`

```go
// precached byte slice reflect type for comparison
var byteSliceType = reflect.TypeOf([]byte(nil))

// ...
func (v *validator) validateAny(value reflect.Value, path string) {
// ...
	case reflect.Slice:
		if value.Type() == byteSliceType {
			// Skip byte slice types since there is no validation on content
			// inside the byte slice.
			return
		}
// ...
```

This method I think is a little easier to read, and allocates the same memory. Over several benchmark runs it seems to be about 4000 ns/op faster, but that could easily be just my system.