@rhysh 2016-02-05
Yes, that's a lot easier to read. Thanks!