@jasdel 2016-02-05
Looks good thanks for taking the time to create the PR. This will improve the resource usage of the SDK on any API operation with []byte as input parameters.